NumericAndStringConverter for GroupingInfo -> GroupInterval
