+++ net/DevExtreme.AspNet.Data.Tests/DataSourceLoadOptionsParserTests.cs
@@ -59,6 +59,21 @@ namespace DevExtreme.AspNet.Data.Tests {
             Assert.IsType<string>(opts.Filter[1]);
         }
 
+        [Fact]
+        public void MustParseNumericAsString() {
+            var opts = new SampleLoadOptions();
+
+            DataSourceLoadOptionsParser.Parse(opts, key => {
+                if(key == DataSourceLoadOptionsParser.KEY_GROUP)
+                    return @"[{""selector"":""freight"",""groupInterval"":100,""isExpanded"":false}]";
+                return null;
+            });
+
+            Assert.Equal("freight", opts.Group[0].Selector);
+            Assert.Equal("100", opts.Group[0].GroupInterval);
+            Assert.False(opts.Group[0].IsExpanded);
+        }
+
     }
 
 }
